refactor to use single responsibility Principle?
