+++ Gemfile
@@ -1,3 +1,3 @@
 source 'https://rubygems.org'
-gem 'rubocop', '>= 1.0', '< 2.0'
 gem 'pry'
+gem 'rubocop', '>= 1.0', '< 2.0'
