+++ app.rb
@@ -1,8 +1,7 @@
-require './book'
-require './rental'
-require './student'
-require './teacher'
-require 'pry'
+require './classes/book'
+require './classes/rental'
+require './classes/student'
+require './classes/teacher'
 
 class App
   def initialize
@@ -47,7 +46,7 @@ class App
 
     print 'Age: '
     age = gets.chomp
-    binding.pry
+
     print 'Student Class: '
     classroom = gets.chomp
 
@@ -109,44 +108,6 @@ class App
     puts 'Rentals:'
     @rentals.each do |rental|
       puts "Date: #{rental.date}, Book \"#{rental.book.title}\" by #{rental.book.author}" if rental.person.id == id
-    end
-  end
-
-  def dashboard
-    puts '----- Choose An Option -----'
-    puts '1 - List all books'
-    puts '2 - List all people'
-    puts '3 - List all rentals for a given person id'
-    puts '4 - Create a book'
-    puts '5 - Create a person'
-    puts '6 - Create a rental'
-    puts '7 - Exit'
-    puts '---------------------------'
-  end
-
-  def option1(input)
-    case input
-    when '1'
-      list_books
-    when '2'
-      list_people
-    when '3'
-      list_rentals_for_person_id
-    end
-  end
-
-  def option2(input)
-    case input
-    when '4'
-      create_book
-    when '5'
-      create_person
-    when '6'
-      create_rental
-    when '7'
-      puts 'Thank you for using this app!'
-    else
-      puts 'Invalid option'
     end
   end
