+++ classes/book.rb
