+++ interfaces/nameable.rb
